requirements.txt
# 데이터베이스
psycopg2-binary>=2.9.0

# 환경 설정
python-dotenv>=1.0.0

# 임베딩 및 ML
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0

# 데이터 처리
pandas>=2.0.0

# HTTP 요청 (Ollama 연결용)
requests>=2.31.0

# 유틸리티
tqdm>=4.65.0